Move new exceptions to Sdk namespace
