+++ Source/xunit.should/Sdk/InRangeException.cs
@@ -1,0 +1,8 @@
+﻿namespace Xunit.Should.Sdk
+{
+    public class InRangeException : RangeException
+    {
+        public InRangeException(object actual, object low, bool lowInclusive, object high, bool highInclusive)
+            : base(actual, low, lowInclusive, high, highInclusive, "Assert.InRange() Failure") { }
+    }
+}
