+++ Source/xunit.should/Sdk/NotInRangeException.cs
@@ -1,0 +1,8 @@
+﻿namespace Xunit.Should.Sdk
+{
+    public class NotInRangeException : RangeException
+    {
+        public NotInRangeException(object actual, object low, bool lowInclusive, object high, bool highInclusive)
+            : base(actual, low, lowInclusive, high, highInclusive, "Assert.NotInRange() Failure") { }
+    }
+}
