+++ Source/xunit.should/xunit.should.csproj
@@ -63,9 +63,12 @@
     <Compile Include="BooleanExtensions.cs" />
     <Compile Include="IComparableExtensions.cs" />
     <Compile Include="IEnumerableExtensions.cs" />
+    <Compile Include="Sdk\InRangeException.cs" />
+    <Compile Include="Sdk\NotInRangeException.cs" />
     <Compile Include="ObjectExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Legacy\ShouldExtensions.cs" />
+    <Compile Include="Sdk\RangeException.cs" />
     <Compile Include="StringComparisonExtensions.cs" />
     <Compile Include="StringExtensions.cs" />
     <Compile Include="ThrowsExtensions.cs" />
